requirements.txt
pandas
ollama
openai
eventregistry
boto3
python-dotenv
